requests-service/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
celery==5.4.0
requests==2.32.3
psycopg2-binary==2.9.9
gunicorn==22.0.0